backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
sqlmodel==0.0.21
pydantic-settings==2.2.1
httpx==0.27.0
pytest==8.3.3
jinja2==3.1.4
python-multipart==0.0.9
